requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
mediapipe>=0.10.0
requests>=2.31.0
pyttsx3>=2.90
pyzmq>=25.0.0
scipy>=1.10.0
matplotlib>=3.7.0
pyyaml>=6.0
comtypes; sys_platform == 'win32'

# pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple